POST /LetterPairTable を別ファイルに切り出した
